requirements.txt
# Core dependencies
torch>=2.1.0
torchaudio>=2.1.0
einops>=0.7.0
numpy>=1.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0

# Model utilities
mamba-ssm>=1.2.0
causal-conv1d>=1.1.0

# Training
accelerate>=0.25.0
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Dataset
datasets>=2.15.0
huggingface-hub>=0.20.0
